Native-specific implementation.
